requirements.txt
streamlit==0.63.0
numpy==1.19.0
tensorflow==2.2.0
Pillow==7.2.0
